test-requirements.txt
nose
vcrpy
